refactor: JspView 클래스 구현
