+++ src/main/java/core/mvc/view/JspView.java
@@ -1,0 +1,29 @@
+package core.mvc.view;
+
+import javax.servlet.RequestDispatcher;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.util.Map;
+
+public class JspView implements View {
+    private final String viewName;
+
+    private static final String DEFAULT_REDIRECT_PREFIX = "redirect:";
+
+    public JspView(String viewName) {
+        this.viewName = viewName;
+    }
+
+    @Override
+    public void render(
+            Map<String, ?> model, HttpServletRequest request, HttpServletResponse response
+    ) throws Exception {
+        if (viewName.startsWith(DEFAULT_REDIRECT_PREFIX)) {
+            response.sendRedirect(viewName.substring(DEFAULT_REDIRECT_PREFIX.length()));
+            return;
+        }
+
+        RequestDispatcher rd = request.getRequestDispatcher(viewName);
+        rd.forward(request, response);
+    }
+}
