+++ src/main/java/core/mvc/view/View.java
@@ -1,4 +1,4 @@
-package core.mvc;
+package core.mvc.view;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
